Fix bug when perspective has not been rendered (#6847)
* fix bug when perspective has not been rendered

* add space before brace

+++ panel/models/perspective.ts
@@ -150,7 +150,9 @@ export class PerspectiveView extends HTMLBoxView {
   }
 
   override remove(): void {
-    this.perspective_element.delete(() => this.worker.terminate())
+    if (this.perspective_element) {
+      this.perspective_element.delete(() => this.worker.terminate())
+    }
     super.remove()
   }
